Added Logic to all components
Applies to All:
qtyMax property added to each component object. Matches the variantQty

methods addToCart and removeFromCart - if-statement logic added to both so that items cannot be added beyond the variant qty and cannot be removed beyond the variant qty (from the cart)
computed qtyReached() - returns the updated qtyMax value
It is then added to the Remove from Cart btn. If BOTH the variantQty (which is updated based on what's added/removed from the cart) AND the qtyMax variable equals zero, the item is out of stock and both buttons are disabled. However, if qtyMax is a positive value, that means you can remove items from the cart, so the item is back in stock